requirements.txt
Flask==2.0.1
Flask-WTF==0.15.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
python-dotenv==0.19.0
WTForms==3.0.1
python-docx==1.1.2
docxtpl==0.20.0